requirements.txt
Django==4.2.0
psycopg==3.2.1 
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg
